finish

+++ philo_bonus/srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: ninieddu <ninieddu@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/02 12:16:59 by ninieddu          #+#    #+#             */
-/*   Updated: 2021/11/05 17:13:11 by ninieddu         ###   ########lyon.fr   */
+/*   Updated: 2021/11/06 11:28:52 by ninieddu         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,12 +27,13 @@ void	*ft_monitor(void *phil)
 		if (ms >= philo->args->time_to_die)
 		{
 			sem_post(philo->args->acting);
-			ft_print_status(philo, "is dead.", 0);
+			ft_print_status(philo, "is dead.", 1);
 			sem_post(philo->args->finish);
-			sem_post(philo->args->end);
 			return (NULL);
 		}
 		sem_post(philo->args->acting);
+		if (philo->meals_count == philo->args->nbr_each_must_eat)
+			return (NULL);
 	}
 }
 
@@ -44,9 +45,12 @@ void	*ft_master_monitor(void *argz)
 	i = -1;
 	args = argz;
 	sem_wait(args->finish);
-	args->stop = 1;
-	while (++i < args->nbr_of_philos)
-		kill(args->philos[i].pid, SIGTERM);
+	if (args->stop == 0)
+		args->stop = 1;
+	if (args->stop == 1)
+		while (++i < args->nbr_of_philos)
+			kill(args->philos[i].pid, SIGTERM);
+	sem_post(args->end);
 	return (NULL);
 }
 
@@ -59,14 +63,15 @@ void	ft_clean(t_args *args)
 	sem_wait(args->end);
 	if (args->stop == 0)
 	{
-		sem_wait(args->end);
+		args->stop = 2;
 		while (++i < args->nbr_of_philos)
 			waitpid(args->philos[i].pid, &status, 0);
+		sem_post(args->finish);
 	}
-	sem_close(args->finish);
 	sem_close(args->acting);
 	sem_close(args->forks);
 	sem_close(args->end);
+	sem_close(args->finish);
 	free(args->philos);
 }
